Add orders
Background
==========
I'm adding an `OrdersController` that should list orders and update the
state of an order to `COMPLETE`.

To load the data I'm creating a service that is reading the orders from
a json file. This can be changed to calling an external service to fetch
the json or fetching the records from the database depending on the
business requirements.

To update an order I created a service that updates the json file. Again
that should be updated when the business requirements change.

The heart of the feature is the `CalculateOrderPrice` service which at
the moment is expecting a hash representation of the orders. I was
considering using structs for the individual orders but I didn't want to
write a serializer for the update function. I'd like to know a little
bit more about the business logic before I reach to other abstractions.